// src/types/profileVisibility.ts
export type ProfileVisibilitySettings = {
  showAvatar: boolean;
  showLocation: boolean;
  showBio: boolean;
  showSkills: boolean;
  showLanguages: boolean;
  showWebsite: boolean;
  showGithub: boolean;
  showLinkedin: boolean;
  showPortfolio: boolean;
  showEmployment: boolean;
  showEducation: boolean;
  showAvailability: boolean;
  showLastSeen: boolean;
};